pyproject.toml
[project]
name = "worker-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aio-pika>=9.5.7",
    "docker>=7.1.0",
    "httpx>=0.28.1",
    "pydantic>=2.11.9",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.1.1",
]
